requirements.txt
streamlit==1.31.1
streamlit-chat==0.1.1
google-generativeai>=0.3.0
praw==7.7.1
python-dotenv==1.0.0 